requirements.txt
Django>=4.2,<4.2.1
psycopg2==2.9.6
django-rest-framework
django-cors-headers>=3.14,<4.0.0
requests
BeautifulSoup4
uWSGI>=2.0.19.1,<2.1